Move validation methods.
